test/UI-test/requirements.txt
allure-pytest==2.8.18
allure-python-commons==2.8.18
pytest==5.4.3
pytest-csv==2.0.2
selenium==3.141.0